+++ docs/source/conf.py
@@ -21,7 +21,7 @@ copyright = '2021, Nelson Gonzabato'
 author = 'Nelson Gonzabato'
 
 # The full version, including alpha/beta/rc tags
-release = "0.1.1"
+release = "0.2.0"
 
 source_suffix = ['.rst', '.md']